chore(lib): rename proof.rs to proof_type.rs
