+++ host/src/interfaces.rs
@@ -1,5 +1,5 @@
 use axum::response::IntoResponse;
-use raiko_lib::proof::ProofType;
+use raiko_lib::proof_type::ProofType;
 use raiko_lib::prover::ProverError;
 use raiko_tasks::{TaskManagerError, TaskStatus};
 use tokio::sync::mpsc::error::TrySendError;
