+++ lib/src/proof_type.rs
